+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 - Admin UI usability improvements
+- Add `description_fi` field for `MountType`
 
 ## [1.0.0] - 2020-07-07